add support for action mode key bindings, like screen/tmux
A key can be bound to enter action mode, like ^A.  The next key
pressed will be the actual action taken, or else action mode will be
exited without processing the key as input.

Also add support for sending the literal action mode key binding,
like binding control+a, a to send a literal control+a.